pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "fastapi-shopping"
version = "0.1.0"
description = "A FastAPI shopping application"

[tool.poetry.dependencies]
python = "^3.9"
anyio = "4.8.0"
asgiref = "3.8.1"
bcrypt = "^4.3.0"
click = "8.1.8"
fastapi = "^0.115.0"
fquery = "^0.4"
pyjwt = "^2.1.0"
passlib = "^1.7.4"
python_jose = "^3.4.0"
python_multiplart = "^0.0.20"
greenlet = "3.1.1"
h11 = "0.14.0"
idna = "3.10"
pydantic_settings = "^2.8.1"
sniffio = "1.3.1"
starlette = "^0.46.1"
strip = "^12.0.0"
typing-extensions = "4.12.2"
uvicorn = "0.17.6"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^4.0.0"
black = "^22.3.0"
isort = "^5.11.4"
flake8 = "^4.0.1"
httpx = "^0.28.1"

[tool.isort]
profile = "black"
